Highlight code client-side
