+++ .changeset/thin-files-flow.md
@@ -1,0 +1,5 @@
+---
+'gitbook': patch
+---
+
+Improve performances by highlighting code client-side if the code block is offscreen
