+++ packages/gitbook/src/components/DocumentView/CodeBlock/highlight.test.ts
@@ -1,9 +1,18 @@
+import type { DocumentBlockCode } from '@gitbook/api';
 import { it, expect } from 'bun:test';
 
-import { highlight } from './highlight';
+import { getInlines, highlight } from './highlight';
+
+async function highlightWithInlines(block: DocumentBlockCode) {
+    const inlines = getInlines(block).map((inline, index) => ({
+        ...inline,
+        body: null,
+    }));
+    return highlight(block, inlines);
+}
 
 it('should parse plain code', async () => {
-    const tokens = await highlight({
+    const tokens = await highlightWithInlines({
         object: 'block',
         type: 'code',
         data: {},
@@ -38,32 +47,35 @@ it('should parse plain code', async () => {
 it('should parse different code in parallel', async () => {
     await Promise.all(
         ['shell', 'scss', 'markdown', 'less', 'scss', 'css', 'scss', 'yaml'].map(async (syntax) =>
-            highlight({
-                object: 'block',
-                type: 'code',
-                data: {
-                    syntax: syntax,
-                },
-                nodes: [
-                    {
-                        object: 'block',
-                        type: 'code-line',
-                        data: {},
-                        nodes: [
-                            {
-                                object: 'text',
-                                leaves: [{ object: 'leaf', marks: [], text: 'Hello world' }],
-                            },
-                        ],
+            highlight(
+                {
+                    object: 'block',
+                    type: 'code',
+                    data: {
+                        syntax: syntax,
                     },
-                ],
-            }),
+                    nodes: [
+                        {
+                            object: 'block',
+                            type: 'code-line',
+                            data: {},
+                            nodes: [
+                                {
+                                    object: 'text',
+                                    leaves: [{ object: 'leaf', marks: [], text: 'Hello world' }],
+                                },
+                            ],
+                        },
+                    ],
+                },
+                [],
+            ),
         ),
     );
 });
 
 it('should parse a multilines plain code', async () => {
-    const tokens = await highlight({
+    const tokens = await highlightWithInlines({
         object: 'block',
         type: 'code',
         data: {},
@@ -138,7 +150,7 @@ it('should parse a multilines plain code', async () => {
 });
 
 it('should parse code with an inline on a single line', async () => {
-    const tokens = await highlight({
+    const tokens = await highlightWithInlines({
         object: 'block',
         type: 'code',
         data: {
@@ -192,10 +204,8 @@ it('should parse code with an inline on a single line', async () => {
                     },
                 },
                 {
-                    type: 'inline',
-                    inline: {
-                        type: 'annotation',
-                    },
+                    type: 'annotation',
+                    body: null,
                     children: [
                         {
                             type: 'shiki',
@@ -229,7 +239,7 @@ it('should parse code with an inline on a single line', async () => {
 });
 
 it('should parse code with an inline on a multiple line', async () => {
-    const tokens = await highlight({
+    const tokens = await highlightWithInlines({
         object: 'block',
         type: 'code',
         data: {
@@ -310,10 +320,8 @@ it('should parse code with an inline on a multiple line', async () => {
                     },
                 },
                 {
-                    type: 'inline',
-                    inline: {
-                        type: 'annotation',
-                    },
+                    type: 'annotation',
+                    body: null,
                     children: [
                         {
                             type: 'shiki',
@@ -347,10 +355,8 @@ it('should parse code with an inline on a multiple line', async () => {
                     },
                 },
                 {
-                    type: 'inline',
-                    inline: {
-                        type: 'annotation',
-                    },
+                    type: 'annotation',
+                    body: null,
                     children: [
                         {
                             type: 'shiki',
@@ -384,7 +390,7 @@ it('should parse code with an inline on a multiple line', async () => {
 });
 
 it('should support code token finishing before the end of the annotation', async () => {
-    const tokens = await highlight({
+    const tokens = await highlightWithInlines({
         object: 'block',
         type: 'code',
         isVoid: false,
@@ -470,10 +476,8 @@ it('should support code token finishing before the end of the annotation', async
                     },
                 },
                 {
-                    type: 'inline',
-                    inline: {
-                        type: 'annotation',
-                    },
+                    type: 'annotation',
+                    body: null,
                     children: [
                         {
                             type: 'shiki',
@@ -507,7 +511,7 @@ it('should support code token finishing before the end of the annotation', async
 });
 
 it('should support multiple code tokens in an annotation', async () => {
-    const tokens = await highlight({
+    const tokens = await highlightWithInlines({
         object: 'block',
         type: 'code',
         isVoid: false,
@@ -615,11 +619,8 @@ it('should support multiple code tokens in an annotation', async () => {
                     },
                 },
                 {
-                    type: 'inline',
-                    inline: {
-                        object: 'inline',
-                        type: 'annotation',
-                    },
+                    type: 'annotation',
+                    body: null,
                     children: [
                         {
                             type: 'shiki',
@@ -653,7 +654,7 @@ it('should support multiple code tokens in an annotation', async () => {
 });
 
 it('should handle \\r', async () => {
-    const tokens = await highlight({
+    const tokens = await highlightWithInlines({
         object: 'block',
         type: 'code',
         data: {
