+++ packages/gitbook/types/memoizee.d.ts
@@ -3,3 +3,9 @@ declare module 'memoizee/weak' {
 
     export = memoizee;
 }
+
+declare module 'memoizee' {
+    declare function memoizee<F extends (...args: any[]) => any>(f: F): F;
+
+    export = memoizee;
+}
